requirements.txt
Django==3.2.4
djangorestframework==3.12.4
requests~=2.31.0
pathlib~=1.0.1
APScheduler==3.7.0
pytz~=2023.3.post1
pyTelegramBotAPI==4.9.0